# openapi.yaml
openapi: 3.1.0
info:
  title: ForgeConfig API
  description: ForgeConfig — REST API for managing network device provisioning, configuration, backups, IPAM, and more.
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/devices:
    get:
      tags: [Devices]
      summary: List devices
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Array of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      tags: [Devices]
      summary: Create device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Created device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/devices/{id}:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    get:
      tags: [Devices]
      summary: Get device
      responses:
        "200":
          description: Device details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Devices]
      summary: Update device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Devices]
      summary: Delete device
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Device ID (integer)
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        minimum: 1
        maximum: 1000
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ────────────────────────────────────────────────────
...
    Device:
      type: object
      required: [id, ip, hostname, config_template, status, device_type, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        mac:
          type: string
        ip:
          type: string
        hostname:
          type: string
        vendor:
          type: string
        model:
          type: string
        serial_number:
          type: string
        config_template:
          type: string
        ssh_user:
          type: string
        ssh_pass:
          type: string
        topology_id:
          type: integer
          format: int64
        topology_role:
          type: string
        hall_id:
          type: integer
          format: int64
        row_id:
          type: integer
          format: int64
        rack_id:
          type: integer
          format: int64
        rack_position:
          type: integer
        device_type:
          type: string
          enum: [internal, external, host]
        status:
          $ref: "#/components/schemas/DeviceStatus"
        last_seen:
          type: string
          format: date-time
        last_backup:
          type: string
          format: date-time
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DeviceStatus:
      type: string
      enum: [online, offline, provisioning, unknown]
    CreateDeviceRequest:
      type: object
      required: [ip, hostname]
      properties:
        mac:
          type: string
        ip:
          type: string
        hostname:
          type: string
        vendor:
          type: string
        model:
          type: string
        serial_number:
          type: string
        config_template:
          type: string
        ssh_user:
          type: string
        ssh_pass:
          type: string
        topology_id:
          type: integer
          format: int64
        topology_role:
          type: string
        device_type:
          type: string
          enum: [internal, external, host]
    UpdateDeviceRequest:
      type: object
      required: [ip, hostname]
      properties:
        mac:
          type: string
        ip:
          type: string
        hostname:
          type: string
        vendor:
          type: string
        model:
          type: string
        serial_number:
          type: string
        config_template:
          type: string
        ssh_user:
          type: string
        ssh_pass:
          type: string
        topology_id:
          type: integer
          format: int64
        topology_role:
          type: string
        device_type:
          type: string
          enum: [internal, external, host]
...